Use ints for some z-levels of vehicles
(openttd r26888) -Codechange: use ints for some z-levels of vehicles
(ic111)

+++ src/roadveh_cmd.cpp
@@ -957,7 +957,7 @@ static void RoadVehCheckOvertake(RoadVehicle *v, RoadVehicle *u)
 	v->overtaking = 1;
 }
 
-static void RoadZPosAffectSpeed(RoadVehicle *v, byte old_z)
+static void RoadZPosAffectSpeed(RoadVehicle *v, int old_z)
 {
 	if (old_z == v->z_pos || _settings_game.vehicle.roadveh_acceleration_model != AM_ORIGINAL) return;
